requirements.txt
httpx>=0.27.0
vllm>=0.5.0
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp>=1.25.0

